Add support for KO error messages
The Gatling Highcharts bundle log is able to include an extra message
about the failure cause in the case of KO responses. This allows the
bundle to group failures by message, so you can see there were X
timeouts, Y 401s etc.

`clj-gatling` already passes back string errors under the `exception`
key, so we just need to make sure to include that in the csv result if
it exists.